% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{AIC_from_data}
\alias{AIC_from_data}
\title{Title}
\usage{
AIC_from_data(
  general_fit_coeffs,
  data,
  dose_var = "dose",
  yield_var = "yield",
  fit_link = "identity"
)
}
\arguments{
\item{fit_link}{}
}
\value{

}
\description{
Title
}
