% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{bar_plots}
\alias{bar_plots}
\title{Bar plots dosimetry}
\usage{
bar_plots(dat, curve, place)
}
\arguments{
\item{dat}{data frame of data values.}

\item{curve}{manual or auto.}

\item{place}{UI or save.}
}
\value{
plot
}
\description{
Bar plots dosimetry
}
