% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodosetools-package.R
\docType{package}
\name{biodosetools-package}
\alias{biodosetools}
\alias{biodosetools-package}
\title{\code{biodosetools} package}
\description{
Shiny App To Be Used By Biological Dosimetry Laboratories
}
\details{
See the README on
\href{https://github.com/biodosetools-team/biodosetools/}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biodosetools-team.github.io/biodosetools/}
  \item \url{https://github.com/biodosetools-team/biodosetools/}
  \item Report bugs at \url{https://github.com/biodosetools-team/biodosetools/issues/}
}

}
\author{
\strong{Maintainer}: Anna Frances-Abellan \email{anna.frances@uab.cat} (\href{https://orcid.org/0000-0003-4738-1712}{ORCID})

Authors:
\itemize{
  \item Alfredo Hernández \email{aldomann.designs@gmail.com} (\href{https://orcid.org/0000-0002-2660-4545}{ORCID})
  \item David Endesfelder \email{dendesfelder@bfs.de}
  \item Pere Puig \email{ppuig@mat.uab.cat} (\href{https://orcid.org/0000-0002-6607-9642}{ORCID})
}

}
\keyword{internal}
