% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodosetools-package.R
\docType{package}
\name{biodosetools}
\alias{biodosetools}
\title{\code{biodosetools} package}
\description{
Shiny App To Be Used By Biological Dosimetry Laboratories
}
\details{
See the README on
\href{https://github.com/biodosetools-team/biodosetools/}{GitHub}
}
