% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_theming.R
\name{bs4DashButtonsStatus}
\alias{bs4DashButtonsStatus}
\title{Creates a custom theme object for a shinydashboard application}
\usage{
bs4DashButtonsStatus(
  home_color,
  home_color_hover,
  home_color_border,
  options_color,
  options_color_hover,
  options_color_border,
  inputs_color,
  inputs_color_hover,
  inputs_color_border,
  results_color,
  results_color_hover,
  results_color_border,
  export_color,
  export_color_hover,
  export_color_border
)
}
\value{
HTML code. Theme for a shinydashboard application.
}
\description{
Creates a custom theme object for a shinydashboard application
}
