% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_theming.R
\name{bs4DashCardsStatus}
\alias{bs4DashCardsStatus}
\title{Creates a custom theme object for a shinydashboard application}
\usage{
bs4DashCardsStatus(options_color, inputs_color, results_color, export_color)
}
\value{
HTML code. Theme for a shinydashboard application.
}
\description{
Creates a custom theme object for a shinydashboard application
}
