% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_theming.R
\name{bs4DashSidebarColor}
\alias{bs4DashSidebarColor}
\title{Creates a custom theme object for a shinydashboard application}
\usage{
bs4DashSidebarColor(back_color)
}
\value{
HTML code. Theme for a shinydashboard application.
}
\description{
Creates a custom theme object for a shinydashboard application
}
