% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_estimation.R
\name{calculate_aberr_IRR}
\alias{calculate_aberr_IRR}
\title{Calculate IRR (this is the same as the Odds-Ratio calculation in IAEA2011)}
\usage{
calculate_aberr_IRR(case_data, fit_coeffs, badge_dose)
}
\arguments{
\item{case_data}{Case data in data frame form.}

\item{fit_coeffs}{Fitting coefficients matrix.}

\item{badge_dose}{Suspected (e.g. badge) dose}
}
\value{
character variable representing IRR in the for (P(X=k|D=0 Gy):P(X=k|D=badge_dose)).
}
\description{
Calculate IRR (this is the same as the Odds-Ratio calculation in IAEA2011)
}
