% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_characteristic_limits.R
\name{calculate_characteristic_limits}
\alias{calculate_characteristic_limits}
\title{Calculate characteristic limits}
\usage{
calculate_characteristic_limits(mu0, n1, alpha = 0.05, beta = 0.1)
}
\arguments{
\item{mu0}{Background rate}

\item{n1}{Number of cells that will be analysed.}

\item{alpha}{Type I error rate, 0.05 by default.}

\item{beta}{Type II error rate, 0.1 by default.}
}
\value{
List of characteristic limits (\code{decision_threshold}, \code{detection_limit}).
}
\description{
Calculate characteristic limits
}
