% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_decision_threshold.R
\name{calculate_decision_threshold}
\alias{calculate_decision_threshold}
\title{Calculate decision thresholds}
\arguments{
\item{fit_results_list}{List of fit results.}

\item{cells}{Number of cells \code{N}.}

\item{count_data}{Count data in data frame form.}

\item{model_formula}{Model formula.}

\item{model_family}{Model family.}

\item{frequency_select}{Whether to use measured frequency or full genome frequency.}

\item{conf_int}{Confidence interval.}

\item{aberr_module}{Aberration module.}
}
\description{
Calculate decision thresholds
}
