% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_limits.R
\name{calculate_table}
\alias{calculate_table}
\title{Calculate the characteristic limits table}
\usage{
calculate_table(input, project_yield, fit_coeffs)
}
\arguments{
\item{input}{all input UI parameters}

\item{project_yield}{function for calculating yield, in utils_estimation.R}

\item{fit_coeffs}{curve coefficients}
}
\value{
list with all the table columns.
}
\description{
Calculate the characteristic limits table
}
