% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_translocations.R
\name{calculate_trans_rate_manual}
\alias{calculate_trans_rate_manual}
\title{Calculate manual translocation rate}
\usage{
calculate_trans_rate_manual(cells, genome_fraction, expected_aberr_value)
}
\arguments{
\item{cells}{Number of cells \code{N}}

\item{genome_fraction}{Genomic fraction used in translocations}

\item{expected_aberr_value}{Expected aberrations}
}
\value{
Translocation rate.
}
\description{
Calculate manual translocation rate
}
