% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{calculate_yield}
\alias{calculate_yield}
\title{Title}
\usage{
calculate_yield(
  dose,
  type = "estimate",
  general_fit_coeffs,
  general_var_cov_mat = NULL,
  protracted_g_value,
  conf_int = 0.95
)
}
\arguments{
\item{dose}{Dose}

\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper"}

\item{conf_int}{}
}
\value{

}
\description{
Title
}
