% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{correct_boundary}
\alias{correct_boundary}
\title{Correct boundary of irradiated fractions to be bounded by 0 and 1}
\usage{
correct_boundary(x)
}
\arguments{
\item{x}{Numeric value.}
}
\value{
Numeric value in [0, 1] range.
}
\description{
Correct boundary of irradiated fractions to be bounded by 0 and 1
}
