% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{correct_conf_int}
\alias{correct_conf_int}
\title{Title}
\usage{
correct_conf_int(
  conf_int,
  general_var_cov_mat,
  protracted_g_value,
  type,
  dose = seq(0, 10, 0.2)
)
}
\arguments{
\item{conf_int}{Confidence interval}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{protracted_g_value}{Protracted G(x) value}

\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper"}

\item{dose}{Dose}
}
\value{

}
\description{
Title
}
