% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{correct_negative_vals}
\alias{correct_negative_vals}
\title{Correct negative values}
\usage{
correct_negative_vals(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
Numeric value corrected to zero if negative
}
\description{
Correct negative values
}
