% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimate.R
\name{correct_yield}
\alias{correct_yield}
\title{Correct yields if they are below the curve}
\usage{
correct_yield(
  yield,
  type = "estimate",
  general_fit_coeffs,
  general_var_cov_mat,
  conf_int
)
}
\arguments{
\item{yield}{Yield}

\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper"}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{conf_int}{Confidence interval}
}
\value{
Numeric value of corrected yield
}
\description{
Correct yields if they are below the curve
}
