% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{DI_plot}
\alias{DI_plot}
\title{Dispersion index}
\usage{
DI_plot(dat, place)
}
\arguments{
\item{dat}{data frame of data values.}

\item{place}{UI or save.}
}
\value{
plot
}
\description{
Dispersion index
}
