% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_content_fractions}
\alias{dna_content_fractions}
\title{DNA Content Fractions of Human Chromosomes}
\format{
A data frame with 24 rows and 3 variables:
\describe{
  \item{chromosome}{Chromosome.}
  \item{fraction_male}{Normalised content of megabases on male human DNA.}
  \item{fraction_female}{Normalised content of megabases on female human DNA.}
}
}
\usage{
dna_content_fractions
}
\description{
Normalised DNA Content of Human Chromosomes.
}
\keyword{datasets}
