% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_content_fractions_ihgsc}
\alias{dna_content_fractions_ihgsc}
\title{DNA Content Fractions of Human Chromosomes (IHGSC)}
\format{
A data frame with 24 rows and 3 variables:
\describe{
  \item{chromosome}{Chromosome.}
  \item{fraction_male}{Normalised content of megabases on male human DNA.}
  \item{fraction_female}{Normalised content of megabases on female human DNA.}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/grc/human/data}
}
\usage{
dna_content_fractions_ihgsc
}
\description{
Normalised DNA Content of Human Chromosomes from
the International Human Genome Sequencing Consortium.
}
\details{
Last accessed in July 2020.
}
\keyword{datasets}
