% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_content_fractions_morton}
\alias{dna_content_fractions_morton}
\title{DNA Content Fractions of Human Chromosomes (Morton 1991)}
\format{
A data frame with 24 rows and 3 variables:
\describe{
  \item{chromosome}{Chromosome.}
  \item{fraction_male}{Normalised content of megabases on male human DNA.}
  \item{fraction_female}{Normalised content of megabases on female human DNA.}
}
}
\source{
\url{https://doi.org/10.1073/pnas.88.17.7474}
}
\usage{
dna_content_fractions_morton
}
\description{
Normalised DNA Content of Human Chromosomes from
"Parameters of the human genome", N. E. Morton (1991).
}
\keyword{datasets}
