% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{dose_boxplot}
\alias{dose_boxplot}
\title{Boxplot dose estimates}
\usage{
dose_boxplot(dat, place)
}
\arguments{
\item{dat}{data frame of data values.}

\item{place}{UI or save.}
}
\value{
plot
}
\description{
Boxplot dose estimates
}
