% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mixed.R
\name{dose_estimation_mx}
\alias{dose_estimation_mx}
\title{Prepare data to dose estimate using mixed fields manual}
\usage{
dose_estimation_mx(input, reactive_data, data_type)
}
\arguments{
\item{input}{what you upload in the UI}

\item{reactive_data}{access to the reactive data}

\item{data_type}{gamma or neutron}
}
\value{
data for dose estimation using manual input.
}
\description{
Prepare data to dose estimate using mixed fields manual
}
