% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{estimate_hetero}
\alias{estimate_hetero}
\title{Title}
\usage{
estimate_hetero(
  case_data,
  general_fit_coeffs,
  general_var_cov_mat,
  fraction_coeff,
  conf_int_yield,
  conf_int_curve,
  protracted_g_value,
  input
)
}
\arguments{
\item{case_data}{Case data in data frame form}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{fraction_coeff}{Fraction coefficient, can be "gamma" or "d0" (to be updated)}

\item{conf_int_yield}{Confidence interval of the yield}

\item{conf_int_curve}{Confidence interval of the curve}

\item{protracted_g_value}{Protracted G(x) value}

\item{input}{UI inputs (to be fixed and parametrized)}
}
\value{

}
\description{
Title
}
