% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{estimate_partial_dolphin}
\alias{estimate_partial_dolphin}
\title{Partial-body dose estimation}
\usage{
estimate_partial_dolphin(
  case_data,
  general_fit_coeffs,
  general_var_cov_mat,
  fraction_coeff,
  conf_int,
  protracted_g_value,
  cov = TRUE,
  genome_fraction,
  aberr_module,
  input
)
}
\arguments{
\item{case_data}{Case data in data frame form}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{fraction_coeff}{Fraction coefficient, can be "gamma" or "d0" (to be updated)}

\item{conf_int}{Confidence interval}

\item{protracted_g_value}{Protracted G(x) value}

\item{cov}{Whether the covariances of the regression coefficients should be considered, otherwise only the diagonal of the covariance matrix is used}

\item{genome_fraction}{Genomic fraction used in translocations, else 1}

\item{aberr_module}{Aberration module}

\item{input}{UI inputs (to be fixed and parametrized)}
}
\value{
List containing estimated doses data frame, estimated fraction of irradiated blood data frame, and AIC
}
\description{
Partial-body dose estimation using Dolphin method.
}
