% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_fitting.R
\name{fit}
\alias{fit}
\title{Perform dose-effect fitting algorithm}
\usage{
fit(
  count_data,
  model_formula,
  model_family,
  fit_link = "identity",
  aberr_module,
  algorithm = c("glm", "maxlik")
)
}
\arguments{
\item{count_data}{Count data in data frame form}

\item{model_formula}{Model formula}

\item{model_family}{Model family}

\item{fit_link}{Family link}

\item{aberr_module}{Aberration module}

\item{algorithm}{Optional selection of algorithm to be used, either "glm" or "maxlik". By default, "glm" is used, with "maxlik" as a fallback method.}
}
\value{
List object containing fit results either using GLM or maxLik optimization
}
\description{
Wrapper for fit_glm_method() and fit_maxlik_method() functions.
}
