% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_interlab.R
\name{fun.curve}
\alias{fun.curve}
\title{Calculate curve}
\usage{
fun.curve(d, coef, curve_type)
}
\arguments{
\item{d}{dose}

\item{coef}{curve coefficients}

\item{curve_type}{type of curve: lin_quad, lin}
}
\value{
Numeric value.
}
\description{
Calculate curve
}
