% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gardner_confidence_table}
\alias{gardner_confidence_table}
\title{Gardner's Confidence Intervals Table}
\format{
A data frame with 1201 rows and 3 variables:
\describe{
  \item{s_est}{Nominal value for the Poisson variable.}
  \item{s_low}{Lower confidence interval for the Poisson variable.}
  \item{s_upp}{Upper confidence interval for the Poisson variable.}
}
}
\source{
\doi{10.1093/biomet/46.3-4.441}
}
\usage{
gardner_confidence_table
}
\description{
Confidence intervals for the expectation of a Poisson variable.
}
\keyword{datasets}
