% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mixed.R
\name{generate_plot_and_download}
\alias{generate_plot_and_download}
\title{Plot at the UI and save}
\usage{
generate_plot_and_download(
  data_type,
  reactive_data,
  output,
  input,
  manual,
  dose_plot,
  name_num
)
}
\arguments{
\item{data_type}{gamma or neutron}

\item{reactive_data}{access to the reactive data}

\item{output}{connection with UI}

\item{input}{what you upload in the UI}

\item{manual}{logical, indicates if using manual input or file data.}

\item{dose_plot}{the number of the dose case to plot.}

\item{name_num}{associates a number to the name.}
}
\value{
plot at UI and saving plot server logic
}
\description{
Plot at the UI and save
}
