% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingAuxFunctions.R
\name{get_decision_threshold}
\alias{get_decision_threshold}
\title{Title}
\usage{
get_decision_threshold(
  fit_results_list,
  cells,
  conf_int = 0.95,
  aberr_module,
  input
)
}
\arguments{
\item{fit_results_list}{List of fit results}

\item{cells}{N}

\item{conf_int}{CI}

\item{aberr_module}{Aberration module}

\item{input}{UI input variable}
}
\value{
A vector with aberr_test and dose_est
}
\description{
Title
}
