% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{get_deltamethod_std_err}
\alias{get_deltamethod_std_err}
\title{Get standard errors using delta method}
\usage{
get_deltamethod_std_err(
  fit_is_lq,
  variable = c("dose", "fraction"),
  mean_estimate,
  cov_estimate,
  protracted_g_value = NA,
  d0 = NA
)
}
\arguments{
\item{fit_is_lq}{Whether the fit is linear quadratic (\code{TRUE}) or linear (\code{FALSE})}

\item{variable}{Variable resulting of the transformation $g(X)$}

\item{mean_estimate}{The estimated mean of $X$}

\item{cov_estimate}{The estimated covariance matrix of $X$}

\item{protracted_g_value}{Protracted $G(x)$ value}
}
\value{
A numeric value containing the standard error of the dose estimate
}
\description{
Delta method for approximating the standard error of a transformation $g(X)$ of a random variable $X = (x1, x2, ...)$, given estimates of the mean and covariance matrix of $X$.
}
