% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{get_estimated_dose_curve}
\alias{get_estimated_dose_curve}
\title{Title}
\usage{
get_estimated_dose_curve(
  est_full_doses,
  general_fit_coeffs,
  general_var_cov_mat,
  protracted_g_value,
  conf_int_yield,
  conf_int_curve,
  conf_int_text_whole,
  conf_int_text_partial,
  conf_int_text_hetero,
  aberr_module,
  input
)
}
\arguments{
\item{input}{}
}
\value{

}
\description{
Title
}
