% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fitting.R
\name{get_fit_dose_curve}
\alias{get_fit_dose_curve}
\title{Get fit dose curve}
\usage{
get_fit_dose_curve(fit_results_list, aberr_name)
}
\arguments{
\item{fit_results_list}{List of fit results}

\item{aberr_name}{Name of the aberration to use in the y-axis}
}
\value{
ggplot object
}
\description{
Get fit dose curve
}
