% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\name{get_fit_glm_method}
\alias{get_fit_glm_method}
\title{Title}
\usage{
get_fit_glm_method(
  count_data,
  model_formula,
  model_family,
  fit_link = "identity"
)
}
\arguments{
\item{count_data}{Count data}

\item{model_formula}{Model formula}

\item{model_family}{Model family}

\item{fit_link}{Family link}
}
\value{
Fit using GLM
}
\description{
Title
}
