% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\name{get_fit_results}
\alias{get_fit_results}
\title{Perform fitting algorithm}
\usage{
get_fit_results(
  count_data,
  model_formula,
  model_family,
  fit_link = "identity",
  aberr_module
)
}
\arguments{
\item{count_data}{Count data in data frame form}

\item{model_formula}{Model formula}

\item{model_family}{Model family}

\item{fit_link}{Family link}

\item{aberr_module}{Aberration module}
}
\value{
List object containing fit results either using GLM or maxLik optimization
}
\description{
Wrapper for get_fit_glm_method() and get_fit_maxlik_method() functions.
}
