% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_translocations.R
\name{get_genome_fraction}
\alias{get_genome_fraction}
\title{Get genome fraction}
\usage{
get_genome_fraction(dna_table, chromosome, color, sex)
}
\arguments{
\item{dna_table}{DNA content fractions table.}

\item{chromosome}{Chromosomes.}

\item{color}{Color of the stains.}

\item{sex}{Sex of the individual}
}
\value{
Fraction of genome hybridised.
}
\description{
Get genome fraction
}
