% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\name{get_model_statistics}
\alias{get_model_statistics}
\title{Title}
\usage{
get_model_statistics(
  model_data,
  fit_coeffs_vec,
  glm_results,
  fit_algorithm,
  response = "yield",
  link = "identity",
  type = "theory"
)
}
\arguments{
\item{model_data}{Data of the model}

\item{fit_coeffs_vec}{Vector of fitting coefficients}

\item{glm_results}{Results of glm}

\item{fit_algorithm}{String of the algorithm used}

\item{response}{Type of response}

\item{link}{Fit link}

\item{type}{Theoretical or raw glm model statistics}
}
\value{
Model statistics
}
\description{
Title
}
