% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_translocations.R
\name{get_translocation_rate_manual}
\alias{get_translocation_rate_manual}
\title{Get manual translocation rate}
\usage{
get_translocation_rate_manual(cells, genome_fraction, expected_aberr_value)
}
\arguments{
\item{cells}{Number of cells \code{N}}

\item{genome_fraction}{Genomic fraction used in translocations}

\item{expected_aberr_value}{Expected aberrations}
}
\value{
Translocation rate.
}
\description{
Get manual translocation rate
}
