% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_translocations.R
\name{get_translocation_rate_sigurdson}
\alias{get_translocation_rate_sigurdson}
\title{Get Sigurdson's translocation rate}
\usage{
get_translocation_rate_sigurdson(
  cells,
  genome_fraction,
  age_value,
  sex_bool = FALSE,
  sex_value = "none",
  smoker_bool = FALSE,
  ethnicity_value = "none",
  region_value = "none"
)
}
\arguments{
\item{cells}{Number of cells \code{N}}

\item{genome_fraction}{Genomic fraction used in translocations}

\item{age_value}{Age of the individual}

\item{sex_bool}{If \code{TRUE}, \code{sex_value} will be used}

\item{sex_value}{Sex of the individual, either "male" of "female"}

\item{smoker_bool}{Whether the individual smokes or not}

\item{ethnicity_value}{Ethnicity of the individual}

\item{region_value}{Region of the individual}
}
\value{
Translocation rate.
}
\description{
Get Sigurdson's translocation rate
}
