% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{include_help}
\alias{include_help}
\title{Include Markdown help}
\usage{
include_help(...)
}
\arguments{
\item{...}{Character vector specifying directory and or file to
point to inside the current package.}
}
\description{
Include Markdown help
}
