% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mixed.R
\name{load_manual_data}
\alias{load_manual_data}
\title{Load manual data}
\usage{
load_manual_data(data_type, input, reactive_data, output, session)
}
\arguments{
\item{data_type}{gamma or neutron}

\item{input}{what you upload in the UI}

\item{reactive_data}{access to the reactive data}

\item{output}{output for the UI}

\item{session}{let interaction with the shiny UI}
}
\value{
manual curve information tables.
}
\description{
Load manual data
}
