% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mixed.R
\name{load_rds_data}
\alias{load_rds_data}
\title{Load and show rds files}
\usage{
load_rds_data(input_data, data_type, reactive_data, output, session)
}
\arguments{
\item{input_data}{what you upload as rds file in the UI.}

\item{data_type}{gamma or neutron.}

\item{reactive_data}{access to the reactive data.}

\item{output}{output for the UI.}

\item{session}{let interaction with the shiny UI.}
}
\value{
uploaded rds file's curve information-tables.
}
\description{
Load and show rds files
}
