% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{load_rmd_report}
\alias{load_rmd_report}
\title{Load RMarkdown report}
\usage{
load_rmd_report(...)
}
\arguments{
\item{...}{Character vector specifying directory and or file to
point to inside the current package.}
}
\description{
Load RMarkdown report
}
