% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_interlab.R
\name{M_estimate}
\alias{M_estimate}
\title{Calculate algB}
\usage{
M_estimate(x, iter_loc = 50, iter_scale = 1000)
}
\arguments{
\item{x}{vector of n observations.}

\item{iter_loc}{number of iteration steps for location estimate (default=50).}

\item{iter_scale}{number of iteration steps for scale estimate  (default=1000).}
}
\value{
Numeric value of zscore using algB.
}
\description{
Calculate algB
}
\examples{
X = c(3.65, 2.5, 4.85)

M_estimate(x = as.numeric(X),
           iter_loc=50,
           iter_scale=1000)
}
