% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mixed.R
\name{plot_estimated_dose_curve_mx}
\alias{plot_estimated_dose_curve_mx}
\title{Criticality Plot}
\usage{
plot_estimated_dose_curve_mx(
  name,
  est_doses,
  fit_coeffs,
  fit_var_cov_mat,
  curve_type,
  protracted_g_value = 1,
  conf_int_curve = 0.95,
  place
)
}
\arguments{
\item{name}{the dose to plot.}

\item{est_doses}{List of dose estimations results.}

\item{fit_coeffs}{Fitting coefficients matrix.}

\item{fit_var_cov_mat}{Fitting variance-covariance matrix.}

\item{curve_type}{gamma or neutron.}

\item{protracted_g_value}{Protracted \eqn{G(x)} value.}

\item{conf_int_curve}{Confidence interval of the curve.}

\item{place}{UI, report or save. Where the plot will be displayed.}
}
\value{
ggcurve
}
\description{
Criticality Plot
}
