% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{plot_interlab_deviation}
\alias{plot_interlab_deviation}
\title{Plot deviation}
\usage{
plot_interlab_deviation(zscore, sum_table, place)
}
\arguments{
\item{zscore}{zscore vector to plot for each lab.}

\item{sum_table}{summary table.}

\item{place}{UI or save.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot deviation
}
