% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{plot_interlab_v2}
\alias{plot_interlab_v2}
\title{Plot zscore v2}
\usage{
plot_interlab_v2(zscore, select_method, sum_table, place)
}
\arguments{
\item{zscore}{zscore vector to plot for each lab.}

\item{select_method}{Zscore algorithm.}

\item{sum_table}{summary table.}

\item{place}{UI or save.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot zscore v2
}
