% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{plot_triage}
\alias{plot_triage}
\title{Plot triage}
\usage{
plot_triage(num_cases, est_doses_whole, est_doses_partial, assessment, place)
}
\arguments{
\item{num_cases}{number of cases.}

\item{est_doses_whole}{dose estimation list of cases.}

\item{est_doses_partial}{dose estimation list of cases.}

\item{assessment}{partial or whole.}

\item{place}{UI or save.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot triage
}
