% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{plot_triage_interlab}
\alias{plot_triage_interlab}
\title{Plot triage interlab}
\usage{
plot_triage_interlab(line_triage, sum_table, place)
}
\arguments{
\item{line_triage}{reference value for selected sample.}

\item{sum_table}{summary table.}

\item{place}{UI or save.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot triage interlab
}
