% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{plot_zscore_all}
\alias{plot_zscore_all}
\title{Plot Z-scores all blind samples}
\usage{
plot_zscore_all(zscore, select_method, place)
}
\arguments{
\item{zscore}{zscore vector to plot for each lab.}

\item{select_method}{chosen algorithm for zscore.}

\item{place}{UI or save.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot Z-scores all blind samples
}
