% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{project_yield_upper}
\alias{project_yield_upper}
\title{Title}
\usage{
project_yield_upper(yield, protracted_g_value, conf_int)
}
\arguments{
\item{conf_int}{}
}
\value{

}
\description{
Title
}
