% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{protracted_g_function}
\alias{protracted_g_function}
\title{Title}
\usage{
protracted_g_function(time, time_0)
}
\arguments{
\item{time}{Time over which the irradiation occurred}

\item{time_0}{The mean lifetime of the breaks, which has been shown to be on the order of ~ 2 hours}
}
\value{

}
\description{
Title
}
