% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimate.R
\name{R_factor}
\alias{R_factor}
\title{Calculate R regression confidence factor}
\usage{
R_factor(general_fit_coeffs, conf_int = 0.95)
}
\arguments{
\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{conf_int}{Confidence interval}
}
\value{
R regression confidence factor
}
\description{
Calculate R regression confidence factor depeding on selected confidence interval and type of fit.
}
