% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{u_test_plot}
\alias{u_test_plot}
\title{U-test}
\usage{
u_test_plot(dat, place)
}
\arguments{
\item{dat}{data frame of data values.}

\item{place}{UI or save.}
}
\value{
plot
}
\description{
U-test
}
