% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mixed.R
\name{update_outputs}
\alias{update_outputs}
\title{Update dose estimation box with yield values}
\usage{
update_outputs(num_cases, reactive_data, output, yield_fun, manual, table)
}
\arguments{
\item{num_cases}{number of cases to estimate.}

\item{reactive_data}{to access reactive data.}

\item{output}{output for the UI.}

\item{yield_fun}{function for calculating yield, in utils_estimation.R}

\item{manual}{logical, indicates if using manual input or file data.}

\item{table}{the data input table.}
}
\value{
yield per dose. Numeric
}
\description{
Update dose estimation box with yield values
}
