% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{yield_boxplot}
\alias{yield_boxplot}
\title{Boxplot yield}
\usage{
yield_boxplot(dat, place)
}
\arguments{
\item{dat}{data frame of data values.}

\item{place}{UI or save.}
}
\value{
plot
}
\description{
Boxplot yield
}
