% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{yield_fun}
\alias{yield_fun}
\title{Calculate yield}
\usage{
yield_fun(dose, general_fit_coeffs, protracted_g_value)
}
\arguments{
\item{dose}{Dose}

\item{general_fit_coeffs}{Generalised fit coefficients matrix}

\item{protracted_g_value}{Protracted G(x) value}
}
\value{
Yield
}
\description{
Calculate yield
}
