% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_estimation.R
\name{AIC_from_data}
\alias{AIC_from_data}
\title{Calculate AIC (Akaike's 'An Information Criterion')}
\usage{
AIC_from_data(
  general_fit_coeffs,
  data,
  dose_var = "dose",
  yield_var = "yield",
  fit_link = "identity"
)
}
\arguments{
\item{general_fit_coeffs}{Generalised fit coefficients matrix.}

\item{data}{Data (dose, yield) to calculate AIC from.}

\item{dose_var}{Name of the dose variable (enquoted).}

\item{yield_var}{Name of the yield variable (enquoted).}

\item{fit_link}{A specification for the model link function.}
}
\value{
Numeric value of AIC.
}
\description{
Calculate AIC (Akaike's 'An Information Criterion')
}
