% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_interlab.R
\name{calc.zValue.new}
\alias{calc.zValue.new}
\title{Calculate ZScore}
\usage{
calc.zValue.new(X, type, alg, c)
}
\arguments{
\item{X}{vector of estimated doses.}

\item{type}{dose or freq.}

\item{alg}{algorithm. Choose between algA, algB or QHampel}

\item{c}{value for reference dose.}
}
\value{
Numeric value of zscore.
}
\description{
Calculate ZScore
}
\examples{

calc.zValue.new(X = c(3.65, 2.5, 4.85),
                type = "dose",
                alg = "algA" ,
                c = c(3.5, 3, 4.25))
}
