% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_aberrations.R
\name{calculate_aberr_table}
\alias{calculate_aberr_table}
\title{Calculate aberrations table}
\usage{
calculate_aberr_table(data, type = c("count", "case"), assessment_u = 1)
}
\arguments{
\item{data}{Count or case data.}

\item{type}{Type of input data. Either "count" and "case".}

\item{assessment_u}{Expected u-value of the assessment. For a Poisson distribution this should be unity.}
}
\description{
Calculate aberrations table
}
