% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_translocations.R
\name{calculate_genome_factor}
\alias{calculate_genome_factor}
\title{Calculate genome fraction}
\usage{
calculate_genome_factor(dna_table, chromosome, color, sex)
}
\arguments{
\item{dna_table}{DNA content fractions table.}

\item{chromosome}{Chromosomes.}

\item{color}{Color of the stains.}

\item{sex}{Sex of the individual}
}
\value{
Fraction of genome hybridised.
}
\description{
Calculate genome fraction
}
