% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimate.R
\name{calculate_yield}
\alias{calculate_yield}
\title{Calculate yield from dose}
\usage{
calculate_yield(
  dose,
  type = "estimate",
  general_fit_coeffs,
  general_var_cov_mat = NULL,
  protracted_g_value,
  conf_int = 0.95
)
}
\arguments{
\item{dose}{Dose}

\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper"}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{protracted_g_value}{Protracted G(x) value}

\item{conf_int}{Confidence interval}
}
\value{
Yield
}
\description{
Calculate yield from dose
}
