% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{calculate_yield_infimum}
\alias{calculate_yield_infimum}
\title{Title}
\usage{
calculate_yield_infimum(
  type = "estimate",
  general_fit_coeffs,
  general_var_cov_mat = NULL,
  conf_int = 0.95
)
}
\arguments{
\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper"}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{conf_int}{Confidence interval}
}
\value{

}
\description{
Title
}
