% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{calculate_yield_infimum}
\alias{calculate_yield_infimum}
\title{Calculate theoretical yield infimum}
\usage{
calculate_yield_infimum(
  type = c("estimate", "lower", "upper"),
  general_fit_coeffs,
  general_fit_var_cov_mat = NULL,
  conf_int = 0.95
)
}
\arguments{
\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper".}

\item{general_fit_coeffs}{Generalised fit coefficients matrix.}

\item{general_fit_var_cov_mat}{Generalised variance-covariance matrix.}

\item{conf_int}{Curve confidence interval, 95\% by default.}
}
\value{
Numeric value of yield infimum.
}
\description{
Calculate theoretical yield infimum
}
