% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{correct_conf_int}
\alias{correct_conf_int}
\title{Title}
\usage{
correct_conf_int(
  conf_int,
  general_var_cov_mat,
  protracted_g_value,
  type,
  dose = seq(0, 10, 0.2)
)
}
\arguments{
\item{general_var_cov_mat}{}
}
\value{

}
\description{
Title
}
