% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{correct_conf_int}
\alias{correct_conf_int}
\title{Correct yield confidence interval}
\usage{
correct_conf_int(
  conf_int,
  general_fit_var_cov_mat,
  protracted_g_value = 1,
  type,
  dose = seq(0, 10, 0.2)
)
}
\arguments{
\item{conf_int}{Confidence interval.}

\item{general_fit_var_cov_mat}{Generalised variance-covariance matrix.}

\item{protracted_g_value}{Protracted \eqn{G(x)} value.}

\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper".}

\item{dose}{Numeric value of dose.}
}
\value{
Numeric value of corrected confidence interval.
}
\description{
Correct yield confidence interval if simple method is required.
}
