% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{correct_yield}
\alias{correct_yield}
\title{Correct yields if they are below the curve}
\usage{
correct_yield(
  yield,
  type = "estimate",
  general_fit_coeffs,
  general_fit_var_cov_mat,
  conf_int
)
}
\arguments{
\item{yield}{Numeric value of yield.}

\item{type}{Type of yield calculation. Can be "estimate", "lower", or "upper".}

\item{general_fit_coeffs}{Generalised fit coefficients matrix.}

\item{general_fit_var_cov_mat}{Generalised variance-covariance matrix.}

\item{conf_int}{Curve confidence interval.}
}
\value{
Numeric value of corrected yield.
}
\description{
Correct yields if they are below the curve
}
