% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{curves_plot}
\alias{curves_plot}
\title{Plot curves}
\usage{
curves_plot(dat, curve, curve_type = "lin_quad", place)
}
\arguments{
\item{dat}{data frame of data values.}

\item{curve}{manual or auto.}

\item{curve_type}{lin or lin_quad.}

\item{place}{UI or save.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot curves
}
