% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_content_fractions_morton}
\alias{dna_content_fractions_morton}
\title{DNA Content Fractions of Human Chromosomes (Morton 1991)}
\format{
A data frame with 24 rows and 3 variables:
\describe{
  \item{chromosome}{Chromosome.}
  \item{fraction_male}{Normalised content of megabases on male human DNA.}
  \item{fraction_female}{Normalised content of megabases on female human DNA.}
}
}
\source{
\doi{10.1073/pnas.88.17.7474}
}
\usage{
dna_content_fractions_morton
}
\description{
Normalised DNA Content of Human Chromosomes from
Morton, N. E. (1991). Parameters of the human genome. Proceedings of the National Academy of Sciences, 88(17), 7474-7476.
}
\keyword{datasets}
