% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{estimate_hetero}
\alias{estimate_hetero}
\title{Title}
\usage{
estimate_hetero(
  case_data,
  general_fit_coeffs,
  general_var_cov_mat,
  fraction_coeff,
  conf_int_yield,
  conf_int_curve,
  protracted_g_value,
  input
)
}
\arguments{
\item{input}{}
}
\value{

}
\description{
Title
}
