% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_estimation.R
\name{estimate_hetero}
\alias{estimate_hetero}
\title{Heterogeneous dose estimation}
\usage{
estimate_hetero(
  case_data,
  fit_coeffs,
  fit_var_cov_mat,
  conf_int_yield,
  conf_int_curve,
  protracted_g_value,
  gamma,
  gamma_error
)
}
\arguments{
\item{case_data}{Case data in data frame form}

\item{fit_coeffs}{Fitting coefficients matrix}

\item{fit_var_cov_mat}{Fitting variance-covariance matrix}

\item{conf_int_yield}{Confidence interval of the yield}

\item{conf_int_curve}{Confidence interval of the curve}

\item{protracted_g_value}{Protracted G(x) value}

\item{gamma}{Survival coefficient of irradiated cells}

\item{gamma_error}{Error of the survival coefficient of irradiated cells}
}
\value{
List containing estimated mixing proportions data frame, estimated yields data frame, estimated doses data frame, estimated fraction of irradiated blood data frame, and AIC
}
\description{
Heterogeneous dose estimation.
}
