% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_estimation.R
\name{estimate_partial_dolphin}
\alias{estimate_partial_dolphin}
\title{Partial-body dose estimation}
\usage{
estimate_partial_dolphin(
  case_data,
  fit_coeffs,
  fit_var_cov_mat,
  conf_int = 0.95,
  protracted_g_value,
  cov = TRUE,
  genome_factor = 1,
  aberr_module,
  gamma
)
}
\arguments{
\item{case_data}{Case data in data frame form}

\item{fit_coeffs}{Fitting coefficients matrix}

\item{fit_var_cov_mat}{Fitting variance-covariance matrix}

\item{conf_int}{Confidence interval, 95\% by default}

\item{protracted_g_value}{Protracted G(x) value}

\item{cov}{Whether the covariances of the regression coefficients should be considered, otherwise only the diagonal of the covariance matrix is used}

\item{genome_factor}{Genomic conversion factor used in translocations, else 1}

\item{aberr_module}{Aberration module}

\item{gamma}{Survival coefficient of irradiated cells}
}
\value{
List containing estimated doses data frame, estimated fraction of irradiated blood data frame, and AIC
}
\description{
Partial-body dose estimation using Dolphin method.
}
