% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{estimate_partial_dolphin}
\alias{estimate_partial_dolphin}
\title{Title}
\usage{
estimate_partial_dolphin(
  case_data,
  general_fit_coeffs,
  general_var_cov_mat,
  fraction_coeff,
  conf_int,
  protracted_g_value,
  cov = TRUE
)
}
\arguments{
\item{cov}{}
}
\value{

}
\description{
Title
}
