% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{estimate_whole_body}
\alias{estimate_whole_body}
\title{Title}
\usage{
estimate_whole_body(
  case_data,
  conf_int_yield,
  conf_int_curve,
  protracted_g_value
)
}
\arguments{
\item{protracted_g_value}{}
}
\value{

}
\description{
Title
}
