% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_estimate.R
\name{estimate_whole_body}
\alias{estimate_whole_body}
\title{Whole-body dose estimation}
\usage{
estimate_whole_body(
  case_data,
  general_fit_coeffs,
  general_var_cov_mat,
  conf_int_yield,
  conf_int_curve,
  protracted_g_value,
  genome_fraction = 1,
  aberr_module
)
}
\arguments{
\item{case_data}{Case data in data frame form}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{conf_int_yield}{Confidence interval of the yield}

\item{conf_int_curve}{Confidence interval of the curve}

\item{protracted_g_value}{Protracted G(x) value}

\item{genome_fraction}{Genomic fraction used in translocations, else 1}

\item{aberr_module}{Aberration module Aberration module}
}
\value{
List containing estimated doses data frame and AIC
}
\description{
Whole-body dose estimation using Merkle's method.
}
