% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_estimation.R
\name{estimate_whole_body_delta}
\alias{estimate_whole_body_delta}
\title{Whole-body dose estimation (delta method)}
\usage{
estimate_whole_body_delta(
  case_data,
  fit_coeffs,
  fit_var_cov_mat,
  conf_int = 0.95,
  protracted_g_value,
  cov = TRUE,
  aberr_module
)
}
\arguments{
\item{case_data}{Case data in data frame form}

\item{fit_coeffs}{Fitting coefficients matrix}

\item{fit_var_cov_mat}{Fitting variance-covariance matrix}

\item{conf_int}{Confidence interval, 95\% by default}

\item{protracted_g_value}{Protracted G(x) value}

\item{cov}{Whether the covariances of the regression coefficients should be considered, otherwise only the diagonal of the covariance matrix is used}

\item{aberr_module}{Aberration module}
}
\value{
List containing estimated doses data frame and AIC
}
\description{
Whole-body dose estimation using delta method.
}
