% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{estimate_whole_body_delta}
\alias{estimate_whole_body_delta}
\title{Whole-body dose estimation}
\usage{
estimate_whole_body_delta(
  case_data,
  general_fit_coeffs,
  general_var_cov_mat,
  conf_int,
  protracted_g_value,
  cov = TRUE,
  aberr_module
)
}
\arguments{
\item{case_data}{Case data in data frame form}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{conf_int}{Confidenve interval}

\item{protracted_g_value}{Protracted G(x) value}

\item{cov}{Whether the covariances of the regression coefficients should be considered, otherwise only the diagonal of the covariance matrix is used}

\item{aberr_module}{Aberration module}
}
\value{
List containing estimated doses data frame and AIC
}
\description{
Whole-body dose estimation using Delta method.
}
