% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFunctions.R
\name{estimate_whole_body_delta}
\alias{estimate_whole_body_delta}
\title{Title}
\usage{
estimate_whole_body_delta(
  case_data,
  general_fit_coeffs,
  general_var_cov_mat,
  conf_int,
  protracted_g_value,
  cov = TRUE,
  aberr_module
)
}
\arguments{
\item{aberr_module}{}
}
\value{

}
\description{
Title
}
