% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_mixed.R
\name{fun.estimate.criticality}
\alias{fun.estimate.criticality}
\title{Calculate absorbed dose for mixed fields exposure}
\usage{
fun.estimate.criticality(
  num_cases,
  dics,
  cells,
  coef_gamma,
  cov_gamma,
  coef_neutron,
  cov_neutron,
  ratio,
  p
)
}
\arguments{
\item{num_cases}{number of cases to estimate.}

\item{dics}{dicentrics.}

\item{cells}{total cells.}

\item{coef_gamma}{Coefficients for gamma curve.}

\item{cov_gamma}{Covariance matrix for gamma curve.}

\item{coef_neutron}{Coefficients for neutron curve.}

\item{cov_neutron}{Covariance matrix for neutron curve.}

\item{ratio}{gamma-neutron ratio.}

\item{p}{photon contribution in mixed curve.}
}
\value{
dose estimation
}
\description{
Calculate absorbed dose for mixed fields exposure
}
