% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{get_estimated_dose_curve}
\alias{get_estimated_dose_curve}
\title{Title}
\usage{
get_estimated_dose_curve(
  est_full_doses,
  protracted_g_value,
  conf_int_yield,
  conf_int_curve
)
}
\arguments{
\item{conf_int_curve}{}
}
\value{

}
\description{
Title
}
