% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_aux_functions.R
\name{get_estimated_dose_curve}
\alias{get_estimated_dose_curve}
\title{Get dose estimation curve}
\usage{
get_estimated_dose_curve(
  est_full_doses,
  general_fit_coeffs,
  general_var_cov_mat,
  protracted_g_value,
  conf_int_yield,
  conf_int_curve,
  conf_int_text_whole,
  conf_int_text_partial,
  conf_int_text_hetero,
  aberr_name
)
}
\arguments{
\item{est_full_doses}{Data frame with yields and dose estimations. It requires dose, yield, type, level columns}

\item{general_fit_coeffs}{Generalized fit coefficients matrix}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{protracted_g_value}{Protracted G(x) value}

\item{conf_int_yield}{Confidence interval of the yield}

\item{conf_int_curve}{Confidence interval of the curve}

\item{conf_int_text_whole}{Text to display confidence interval for whole-body estimation}

\item{conf_int_text_partial}{Text to display confidence interval for partial-body estimation}

\item{conf_int_text_hetero}{Text to display confidence interval for heterogeneous estimation}

\item{aberr_name}{Name of the aberration to use in the y-axis}
}
\value{
ggplot object
}
\description{
Get dose estimation curve
}
