% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingAuxFunctions.R
\name{get_fit_dose_curve}
\alias{get_fit_dose_curve}
\title{Title}
\usage{
get_fit_dose_curve(fit_results_list)
}
\arguments{
\item{fit_results_list}{List of fit results}
}
\value{
ggplot object
}
\description{
Title
}
