% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingAuxFunctions.R
\name{get_fit_dose_curve}
\alias{get_fit_dose_curve}
\title{Get fit dose curve}
\usage{
get_fit_dose_curve(fit_results_list, aberr_module, input)
}
\arguments{
\item{fit_results_list}{List of fit results}

\item{aberr_module}{Aberration module}

\item{input}{UI input variable}
}
\value{
ggplot object
}
\description{
Get fit dose curve
}
