% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{get_fit_glm_method}
\alias{get_fit_glm_method}
\title{Perform GLM (Generalized Linear Model) fitting}
\usage{
get_fit_glm_method(
  count_data,
  model_formula,
  model_family,
  fit_link = "identity",
  aberr_module
)
}
\arguments{
\item{count_data}{Count data in data frame form}

\item{model_formula}{Model formula}

\item{model_family}{Model family}

\item{fit_link}{Family link}

\item{aberr_module}{Aberration module}
}
\value{
List object containing GLM fit results
}
\description{
Perform GLM (Generalized Linear Model) fitting
}
