% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{get_fit_maxlik_method}
\alias{get_fit_maxlik_method}
\title{Perform max-likelihood optimization fitting}
\usage{
get_fit_maxlik_method(
  data,
  model_formula,
  model_family = c("automatic", "poisson", "quasipoisson", "nb2"),
  fit_link
)
}
\arguments{
\item{data}{Count data}

\item{model_formula}{Model formula}

\item{model_family}{Model family}

\item{fit_link}{Family link}
}
\value{
List object containing maxLik fit results
}
\description{
Method based on the paper by Oliveira et al. (2011) <doi:10.1007/s00184-009-0295-7>
}
