% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\name{get_fit_results}
\alias{get_fit_results}
\title{Title}
\usage{
get_fit_results(count_data, model_formula, model_family, fit_link = "identity")
}
\arguments{
\item{count_data}{Count data}

\item{model_formula}{Model formula}

\item{model_family}{Model family}

\item{fit_link}{Family link}
}
\value{
Fit results either using GLM or maxLik optimization
}
\description{
Title
}
