% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{get_model_statistics}
\alias{get_model_statistics}
\title{Calculate model statistics}
\usage{
get_model_statistics(
  model_data,
  fit_coeffs_vec,
  glm_results = NULL,
  fit_algorithm = NULL,
  response = "yield",
  link = "identity",
  type = c("theory", "raw"),
  Y = NULL,
  mu = NULL,
  n = NULL,
  npar = NULL,
  genome_fraction = NULL,
  aberr_module
)
}
\arguments{
\item{model_data}{Data of the model}

\item{fit_coeffs_vec}{Vector of fitting coefficients}

\item{glm_results}{Results of glm}

\item{fit_algorithm}{String of the algorithm used}

\item{response}{Type of response}

\item{link}{Fit link}

\item{type}{Theoretical or raw glm model statistics}

\item{Y}{Y response (required in constraint-maxlik-optimization)}

\item{mu}{mu response required in constraint-maxlik-optimization)}

\item{n}{number of parameters (required in constraint-maxlik-optimization)}

\item{npar}{number of parameters (required in constraint-maxlik-optimization)}

\item{genome_fraction}{Genomic fraction used in translocations}

\item{aberr_module}{Aberration module}
}
\value{
Model statistics data frame
}
\description{
Calculate model statistics
}
