% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{plot_estimated_dose_curve}
\alias{plot_estimated_dose_curve}
\title{Plot dose estimation curve}
\usage{
plot_estimated_dose_curve(
  est_doses,
  fit_coeffs,
  fit_var_cov_mat,
  protracted_g_value = 1,
  conf_int_curve,
  aberr_name
)
}
\arguments{
\item{est_doses}{List of dose estimations results from \code{estimate_*()} family of functions.}

\item{fit_coeffs}{Fitting coefficients matrix.}

\item{fit_var_cov_mat}{Fitting variance-covariance matrix.}

\item{protracted_g_value}{Protracted \eqn{G(x)} value.}

\item{conf_int_curve}{Confidence interval of the curve.}

\item{aberr_name}{Name of the aberration to use in the y-axis.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot dose estimation curve
}
