% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fitting.R
\name{plot_fit_dose_curve}
\alias{plot_fit_dose_curve}
\title{Plot fit dose curve}
\usage{
plot_fit_dose_curve(fit_results_list, aberr_name)
}
\arguments{
\item{fit_results_list}{List of fit results.}

\item{aberr_name}{Name of the aberration to use in the y-axis.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot fit dose curve
}
