% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{plot_triage}
\alias{plot_triage}
\title{Plot triage}
\usage{
plot_triage(num_cases, est_doses_whole, est_doses_partial, assessment, place)
}
\arguments{
\item{num_cases}{number of cases.}

\item{est_doses_whole}{dose estimation list of cases.}

\item{est_doses_partial}{dose estimation list of cases.}

\item{assessment}{partial or whole.}

\item{place}{UI or save.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot triage
}
\examples{
est_doses_whole <- data.frame(
    ID = c("example1", "example2"),
    lower = c(1.407921, 2.604542),
    estimate = c(1.931245, 3.301014),
    upper = c(2.632199, 4.221749)
  )


plot_triage(
  num_cases = 2,
  est_doses_whole ,
  est_doses_partial = NULL,
  assessment = "whole",
  place = "UI"
)

}
