% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\name{prepare_maxlik_count_data}
\alias{prepare_maxlik_count_data}
\title{Title}
\usage{
prepare_maxlik_count_data(count_data, model_formula, aberr_module)
}
\arguments{
\item{count_data}{Count data in dataframe form}

\item{model_formula}{Formula}

\item{aberr_module}{Aberration module}
}
\value{
Parsed count data
}
\description{
Title
}
