% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_fitting.R
\name{prepare_maxlik_count_data}
\alias{prepare_maxlik_count_data}
\title{Prepare count data for max-likelihood optimization fitting}
\usage{
prepare_maxlik_count_data(
  count_data,
  model_formula,
  aberr_module = c("dicentrics", "translocations", "micronuclei")
)
}
\arguments{
\item{count_data}{Count data in data frame form.}

\item{model_formula}{Model formula.}

\item{aberr_module}{Aberration module.}
}
\value{
Data frame of parsed count data.
}
\description{
Prepare count data for max-likelihood optimization fitting
}
