% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{project_yield}
\alias{project_yield}
\title{Title}
\usage{
project_yield(
  yield,
  type = "estimate",
  general_fit_coeffs,
  general_var_cov_mat = NULL,
  protracted_g_value,
  conf_int = 0.95
)
}
\arguments{
\item{conf_int}{}
}
\value{

}
\description{
Title
}
