% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{protracted_g_function}
\alias{protracted_g_function}
\title{Calculate protracted function \eqn{G(x)}}
\usage{
protracted_g_function(time, time_0 = 2)
}
\arguments{
\item{time}{Time over which the irradiation occurred.}

\item{time_0}{The mean lifetime of the breaks, which has been shown to be on the order of ~ 2 hours (default value).}
}
\value{
Numeric value of \eqn{G(x)}.
}
\description{
Calculation based on the paper by Lea, D. E. & Catcheside, D. G. (1942). The
mechanism of the induction by radiation of chromosome aberrations
in\emph{Tradescantia}. Journal of Genetics, 44(2-3), 216-245.
<doi:10.1007/BF02982830>.
}
