% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_interlab.R
\name{QHampel}
\alias{QHampel}
\title{Calculate QHampel}
\usage{
QHampel(y, lab, tol.G1 = 1e-06)
}
\arguments{
\item{y}{vector of data values.}

\item{lab}{corresponding lab numbers.}

\item{tol.G1}{decimal place accuracy of the numerator of s.star.}
}
\value{
Numeric value of zscore using QHampel algorithm.
}
\description{
Calculate QHampel
}
