% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_interlab.R
\name{summary_curve_tables}
\alias{summary_curve_tables}
\title{Summary and curve tables}
\usage{
summary_curve_tables(num_labs, list_lab_names, all_rds)
}
\arguments{
\item{num_labs}{number of laboratories participating.}

\item{list_lab_names}{list with the laboratory names.}

\item{all_rds}{list with the rds files.}
}
\value{
list(sorted_table_ilc, sorted_table_curve)
}
\description{
Summary and curve tables
}
