% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAuxFunctions.R
\name{yield_error_fun}
\alias{yield_error_fun}
\title{Title}
\usage{
yield_error_fun(dose, general_var_cov_mat = NULL, protracted_g_value)
}
\arguments{
\item{dose}{Dose}

\item{general_var_cov_mat}{Generalized variance-covariance matrix}

\item{protracted_g_value}{Protracted G(x) value}
}
\value{

}
\description{
Title
}
