% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_estimation.R
\name{yield_error_fun}
\alias{yield_error_fun}
\title{Calculate yield error}
\usage{
yield_error_fun(dose, general_fit_var_cov_mat = NULL, protracted_g_value = 1)
}
\arguments{
\item{dose}{Numeric value of dose.}

\item{general_fit_var_cov_mat}{Generalised variance-covariance matrix.}

\item{protracted_g_value}{Protracted \eqn{G(x)} value.}
}
\value{
Numeric value of yield error.
}
\description{
Calculate yield error using Merkle's method
}
